backend/requirements.txt
# Core Django
Django==4.2.7

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Authentication
dj-rest-auth==5.0.2
django-allauth==0.58.2

# CORS
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# File handling
Pillow==10.1.0

# Environment variables
python-dotenv==1.0.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# API Documentation
drf-yasg==1.21.7

# Filtering
django-filter==23.5

# Validation
drf-spectacular==0.27.1

# Security
python-jose==3.3.0
cryptography==42.0.5

# Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
ipython==8.18.1
django-debug-toolbar==4.2.0

# Resume Parser Dependencies
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
nltk==3.8.1
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
